Unify the operations handling.
